requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
beautifulsoup4>=4.12.0
pypdf>=3.10.0
pytesseract>=0.4.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-genai>=0.1.0
playwright>=1.35.0
pytest>=7.0.0
